requirements.txt
streamlit 
pillow
easyocr
transformers
torch